Saved changes

+++ src/app/admin/distributor/distributortargetvideo/distributortargetvideolist/distributortargetvideolist.component.ts
@@ -11,6 +11,8 @@ declare var $: any;
 export class DistributortargetvideolistComponent implements OnInit {
   p:number=1;
   alllist: any;
+  ngxService: any;
+  toastr: any;
   constructor(public distributorService: DistributorService, public router: Router) { }
 
   ngOnInit() {
@@ -18,28 +20,70 @@ export class DistributortargetvideolistComponent implements OnInit {
       this.alllist = list['data'];
     });
 
-    $(document).ready(function () {
-      setTimeout(() => {
-        let table = $('#pagedatatable').DataTable({
-          ordering: true,
-          lengthChange: false,
-          showNEntries: false,
-        })
-      }, 4000)
-    })
+    // $(document).ready(function () {
+    //   setTimeout(() => {
+    //     let table = $('#pagedatatable').DataTable({
+    //       ordering: true,
+    //       lengthChange: false,
+    //       showNEntries: false,
+    //     })
+    //   }, 4000)
+    // })
   }
 
-  delete(id) {
+  // delete(id) {
+  //   var obj = {
+  //     video_id: id
+  //   };
+
+  //   this.distributorService.deleteDistirbutorVideoById(id).subscribe(res => {
+  //     if (res['result'] == true) {
+  //       location.reload();
+  //       // this.router.navigate(['/admin', 'distributorvideo-list']);
+  //     }
+  //   });
+  // }
+  deleteItem(event) {
     var obj = {
-      video_id: id
-    };
-
-    this.distributorService.deleteDistirbutorVideoById(obj).subscribe(res => {
-      if (res['result'] == true) {
-        this.router.navigate(['/admin', 'distributorvideo-list']);
-      }
-    });
+      video_id: event
+        };
+    
+    let ans = window.confirm('Do you really want to delete this target  video?');
+    if (ans) {
+      this.distributorService.deleteDistirbutorVideoById(obj).subscribe(res => {
+        if (res['result'] == true) {
+          alert(`data deleted`);
+          location.reload();
+         
+        } else {
+          alert(`data not deleted`);
+          // Your code for the default else part goes here
+          // For example, you can add a function call or other logic
+          // this.handleDeleteError();
+        }
+      });
+    }
   }
+
+
+
+  // deleteItem(id) {
+  //   var obj = {
+  //     id: id
+  //   };
+  //   // this.ngxService.start();
+  //   this.distributorService.deleteDistirbutorVideoById(obj).subscribe(res => {
+  //     if (res['result'] == true) {
+  //       this.toastr.success("Video deleted successfully!");
+  //       this.router.navigate(['/admin', 'redirectself'], { state: ['/admin', 'distributorvideoall-list'] });
+  //     }
+  //     if (res['error'] == true) {
+  //       this.toastr.error("Something went wrong " + res['message']);
+  //     }
+  //   });
+  //   // this.ngxService.stop();
+  // }
+
 
   getForEdit(event) {
     var obj = {
